Refactor: Generate Models.py to Folder Models
