+++ img2xl_app/models/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+import UserProfile
+import UploadedFile
+import ExtractedResult
+import UsageLog
+import Notification
+
